+++ Dfe.PrepareConversions/Dfe.PrepareConversions.Data/Services/Interfaces/ISignificantChangeProjectRepository.cs
@@ -32,6 +32,7 @@ public interface ISignificantChangeProjectRepository
    Task SetStakeholderConsultation(int id, SetSignificantChangeStakeholderConsultationCommand command);
    Task SetEqualitiesImpactAssessment(int id, SetSignificantChangeEqualitiesImpactAssessmentCommand command);
    Task SetReligiousBodyConsultation(int id, SetSignificantChangeReligiousBodyConsultationCommand command);
+   Task SetLocalAuthorityObjections(int id, SetSignificantChangeLocalAuthorityObjectionsCommand command);
    Task SetProjectDates(int id, SetSignificantChangeProjectDatesCommand command);
    Task SetStakeholderObjections(int id, SetSignificantChangeStakeholderObjectionsCommand command);
    Task SetAdmissionVariationConsultation(int id, SetSignificantChangeAdmissionVariationConsultationCommand command);